+++ CHANGELOG.md
@@ -16,6 +16,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Mode-specific bindings can now be bound in any mode for easier macros
 - `--help` output is more compact now and uses more neutral palette
 
+### Fixed
+
+- Hyperlink preview not being shown when the terminal has exactly 2 lines
+
 ## 0.12.1
 
 ### Fixed