+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Warnings for unused configuration file options
+- Config option `persist` in `hints` config section
 
 ### Changed